rename : useApis -> user/getUserInform로 변경
