+++ src/pages/home/components/post/PostItem.tsx
@@ -4,7 +4,7 @@ import { useNavigate, Link } from 'react-router-dom'
 import tw, { styled } from 'twin.macro'
 import { useProfileStore } from '@/stores/userProfileStore'
 import NoProfileThumbnailIcon from '@/pages/search/components/NoProfileThumbnailIcon'
-import { getUserInform } from '@/apis/userApis'
+import { getUserInform } from '@/apis/user/getUserInform'
 import CommentIcon from '@/assets/icons/comment.svg'
 import HeartIcon from '@/assets/icons/following.svg'
 import NoImage from '@/assets/temp/noImage.png'
